requirements.txt
CoffeeScript==1.0.5
Fabric==1.8.2
Flask==0.10.1
Flask-OpenID==1.1.1
Pygments==1.6
chaussette==1.0
circus==0.10.0
gevent==1.0
gevent-websocket==0.9.3
prat-misaka==1.1.5
pymongo==2.5.2
pyzmq==14.1.0
redis==2.9.1
toml==0.8.1
click==3.3

# ancillary packages for gevent-websocket
ujson==1.33
wsaccel==0.6.2
